bug-hunt #85: -Wduplicated-branches - fix 24 cases of identical if/else branches
Key changes:
- Widen open_flags from uint8_t to uint32_t (process.h) so that
  O_NONBLOCK (04000) and O_ASYNC (0x2000) actually fit in the field
- Fix sys_ioctl.c: remove (uint8_t) casts that truncated constants to 0
  making both if/else branches effectively no-ops
- Fix syscall.c sys_socket/sys_accept: simplify ternary where both branches
  produced identical results via unsigned/signed casts
- Fix syscall.c sys_unlinkat: consolidate copy-paste that called vfs_unlink
  in both AT_REMOVEDIR and non-AT_REMOVEDIR branches
- Fix syscall.c sys_select: merge duplicated FD_CLR branches
- Fix rtl8139.c CAPR calculation: uint16_t cast handles wraparound natively
- Fix igb.c queue count: consolidated identical I350/82576 branches
- Fix acpi_aml.c body_len: PkgLength is same for extended and normal opcodes
- Fix process.c CLONE_FILES: remove empty if/else (both cases are no-ops)
- Fix tmpfs.c error path: unconditional free_node_pages
- Fix ext4.c BGD offset: both block-size cases compute identical result
- Fix ext2.c extent clear: identical branches for extent/indirect inodes
- Fix ext4_extents.c: _Static_assert confirms extent leaf/index are same size
- Fix net/conntrack.c: FIN_WAIT_2 transition converges from both paths
- Fix wireguard.c: remove dead ternary carry; both responder types use
  sender_index
- Fix garp/mrp/cake/eventfd/pthread/debugfs/file_lock/usb_cdc_ether/
  madvise_ext: consolidate various identical ternary/if-else branches
- Add -Wduplicated-branches to CFLAGS permanently

+++ src/kernel/sys_ioctl.c
@@ -72,9 +72,9 @@ static int ioctl_fionbio(struct process *p, int fd, uint64_t arg)
 	if (copy_from_user(&val, arg, sizeof(val)) < 0)
 		return -EFAULT;
 	if (val)
-		p->fd_table[fd].open_flags |= (uint8_t)04000;  /* O_NONBLOCK */
+		p->fd_table[fd].open_flags |= 04000;  /* O_NONBLOCK */
 	else
-		p->fd_table[fd].open_flags &= (uint8_t)~04000;
+		p->fd_table[fd].open_flags &= ~04000;
 	return 0;
 }
 
@@ -88,9 +88,9 @@ static int ioctl_fioasync(struct process *p, int fd, uint64_t arg)
 	if (copy_from_user(&val, arg, sizeof(val)) < 0)
 		return -EFAULT;
 	if (val)
-		p->fd_table[fd].open_flags |= (uint8_t)0x2000;  /* O_ASYNC */
+		p->fd_table[fd].open_flags |= 0x2000;  /* O_ASYNC */
 	else
-		p->fd_table[fd].open_flags &= (uint8_t)~0x2000;
+		p->fd_table[fd].open_flags &= ~0x2000;
 	return 0;
 }
